requirements.txt
# Core dependencies
torch>=2.0.0,<2.5.0
nibabel>=3.0.0,<6.0.0
numpy>=1.20.0,<2.0.0
h5py>=3.0.0,<4.0.0
pyyaml>=5.0.0,<7.0.0

# Optional dependencies for data preparation
# geant4 (requires separate installation)
# penelope (requires separate installation)
